+++ src-tauri/src/lib.rs
@@ -19,6 +19,7 @@ pub fn run() {
     updater_proxy::configure();
 
     tauri::Builder::default()
+        .plugin(tauri_plugin_clipboard_manager::init())
         .plugin(tauri_plugin_fs::init())
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_shell::init())